do the feature proper
